feat(platform): #1264 css changes, code reorganized into proper folder
also,
rtl fixes for icon and text
secondary icon css fix
truncation with inline-flex issue resolved
