--- libs/platform/src/lib/components/menu/menu-comp/menu-item.component.html
@@ -1,81 +1,0 @@
-<ng-container
-    *ngTemplateOutlet="
-        selectable ? selectableItemTemplate : itemTemplate;
-        context: {
-            label: label,
-            i: index,
-            selected: selected,
-            active: _isActive,
-            item: item,
-            group: group,
-            separated: separated,
-            secondaryIcon: secondaryIcon,
-            disabled: disabled,
-            customLabel: customLabel
-        }
-    "
-    [ngClass]="{ 'fd-menu__list--separated': separated }"
->
-</ng-container>
-<ng-template #itemTemplate let-label="label" let-i="i" let-active="active">
-    <li data-tag="menu-item" tabindex="0" [attr.data-index]="i" [attr.disabled]="disabled">
-        <!-- removed from li [ngClass]="{ 'fd-menu__list--separated': separated }" -->
-        <div class="fd-menu__addon-before" *ngIf="icon; else noIcon">
-            <span [ngClass]="[icon]" data-tag="menu-item__icon" [attr.title]="icon">
-                <!-- removed from span , separated ? 'fd-menu__list--separated' : '' -->
-            </span>
-        </div>
-        <span
-            class="fd-menu__item"
-            [ngClass]="[active ? 'is-selected' : '', disabled ? 'is-disabled' : '']"
-            (click)="onItemClick()"
-            role="button"
-            data-tag="menu-item__button"
-            [attr.title]="customLabel ? customLabel : label"
-        >
-            {{ label }}
-        </span>
-        <div class="fd-menu__addon-after" *ngIf="secondaryIcon; else noIcon">
-            <span
-                [ngClass]="[secondaryIcon ? secondaryIcon : '']"
-                data-tag="menu-item__icon"
-                [attr.title]="secondaryIcon"
-            >
-                <!-- removed from span , separated ? 'fd-menu__list--separated' : '' -->
-            </span>
-        </div>
-    </li>
-</ng-template>
-<ng-template #noIcon><span [ngClass]="{ 'fd-menu__list--separated': separated }"></span></ng-template>
-<!-- works fine don't change -->
-<ng-template #selectableItemTemplate let-label="label" let-i="i" let-selected="selected" let-active="active">
-    <li data-tag="menu-item" tabindex="0" [attr.data-index]="i" [attr.disabled]="disabled">
-        <!-- removed from li         [ngClass]="{ 'fd-menu__list--separated': separated }" -->
-        <div class="fd-menu__addon-before">
-            <span
-                *ngIf="selected; else noIcon"
-                [ngClass]="[icon ? icon : 'sap-icon--accept']"
-                data-tag="menu-item__icon"
-                [attr.title]="icon"
-            >
-                <!-- removed from span , separated ? 'fd-menu__list--separated' : '' -->
-            </span>
-        </div>
-        <span
-            class="fd-menu__item"
-            [ngClass]="[active ? 'is-selected' : '', disabled ? 'is-disabled' : '']"
-            (click)="onItemClick()"
-            [attr.aria-selected]="selected ? 'true' : 'false'"
-            role="button"
-            data-tag="menu-item__button"
-            [attr.title]="customLabel ? customLabel : label"
-        >
-            {{ label }}
-        </span>
-    </li>
-</ng-template>
-
-<!-- maybe add 'fd-menu__list--separated': separated to icon's ngclass         [ngClass]="[icon ? icon : 'sap-icon--accept',{'fd-menu__list--separated': separated}]"
- -->
-
-<!-- [ngClass]="[active ? 'is-selected' : '', disabled ? 'is-disabled' : '']" -->
