--- libs/platform/src/lib/components/menu/menu-comp/menu.component.html
@@ -1,83 +1,0 @@
-<div
-    class="fd-menu"
-    data-tag="menu"
-    [ngClass]="{
-        'fd-menu__list--separated': separator,
-        'scrolling-menu': scrolling
-    }"
-    [ngStyle]="{ height: scrollLimit * 40 + 8 + 'px', width: width }"
-    [style.text-align]="textAlign"
->
-    <ng-container *ngFor="let group of groups; let i = index">
-        <ng-container
-            *ngTemplateOutlet="
-                group.label ? listWithGroupTitleTemplate : listWithoutGroupTitleTemplate;
-                context: { group: group, i: i }
-            "
-        >
-        </ng-container>
-    </ng-container>
-</div>
-
-<ng-template #listWithoutGroupTitleTemplate let-group="group" let-i="i">
-    <ul class="fd-menu__list" data-tag="menu__list" [attr.data-index]="i">
-        <!-- 
-          separator class in ul does not make any difference
-          do not add , 'fd-menu--addon-before': isIconPresent() in <ul> it simply does 32px padding  -->
-        <ng-container *ngFor="let item of group.children; let j = index">
-            <!-- making component of {{ item.label }} value of i {{ i }} and j {{ j }} unique index is
-            {{ i * group.children.length + j }} -->
-            <fdp-menu-item
-                [label]="item.label"
-                [selectable]="item.selectable"
-                [selected]="item.selected"
-                [icon]="item.icon || group.icon"
-                [secondaryIcon]="item.secondaryIcon"
-                [item]="item"
-                [group]="group"
-                [index]="j"
-                [separated]="separator"
-                [disabled]="item.disabled"
-                [customLabel]="item.customLabel || group.customLabel"
-                (itemClick)="onItemClick(item, group)"
-                (keyup)="onTargetKeyup($event)"
-            >
-            </fdp-menu-item>
-        </ng-container>
-    </ul>
-</ng-template>
-
-<ng-template #listWithGroupTitleTemplate let-group=" group" let-i="i">
-    <div class="fd-menu__group" data-tag="menu__group" [attr.data-index]="i">
-        <h3
-            class="fd-menu__title"
-            data-tag="menu__group--title"
-            [attr.title]="group.customLabel ? group.customLabel : group.label"
-            tabindex="0"
-        >
-            {{ group.label }}
-        </h3>
-        <ul class="fd-menu__list" data-tag="menu__list">
-            <!-- not having separator here fails the test, but no diff to ui. so fix the test instead -->
-            <ng-container *ngFor="let item of group.children; let j = index">
-                <!-- making component of {{ item.label }} value of i {{ i }} and j {{ j }} unique index is
-                {{ i * group.children.length + j }} -->
-                <fdp-menu-item
-                    [label]="item.label"
-                    [selectable]="item.selectable"
-                    [selected]="item.selected"
-                    [icon]="item.icon || group.icon"
-                    [secondaryIcon]="item.secondaryIcon"
-                    [item]="item"
-                    [group]="group"
-                    [index]="j"
-                    [separated]="separator"
-                    [disabled]="item.disabled"
-                    [customLabel]="item.customLabel || group.customLabel"
-                    (itemClick)="onItemClick(item, group)"
-                    (keyup)="onTargetKeyup($event)"
-                ></fdp-menu-item>
-            </ng-container>
-        </ul>
-    </div>
-</ng-template>
