--- libs/platform/src/lib/components/menu/menu-comp/menu.component.scss
@@ -1,39 +1,0 @@
-/*BEGIN- GENERAL SPECIFICATION should we keep it here or in menu item scss?*/
-
-.fd-menu {
-    //  background-color: #FFFFFF;//background-color is correct in styles itself. REMOVE THIS
-    border-radius: 4px;
-    border: 1px solid #d9d9d9; //need to specify it explicitly
-    box-shadow: 0px 0px 4px 0px rgba(0, 0, 0, 0.05);
-}
-
-// .fd-menu--addon-before .fd-menu__item{  //fd-menu--addon-before represents the container of the menu item
-//   padding-left: 0px;
-// }
-
-// .fd-menu > .fd-menu__list--separated{
-//   color: red;
-// }
-// .fd-menu
-
-/*END- GENERAL SPECIFICATION*/
-
-.scrolling-menu {
-    overflow-y: scroll;
-}
-
-.fd-menu__title {
-    height: 32px;
-    padding: 10px 16px; //32px -12 px font-size = 20px /2 =10px top/bottom
-}
-
-.fd-menu__list--separated>.fd-menu__list > * > *, //without groups
-.fd-menu__list--separated>.fd-menu__group>.fd-menu__list > * > * //with groups
- {
-    border-bottom-style: solid;
-    border-bottom-width: 1px;
-    color: var(--fd-menu-item-separator-color);
-    // &:first-child {
-    //     border-top-width: 0;
-    // }
-}
