+++ libs/platform/src/lib/components/menu/menu-component/menu-item.component.html
@@ -12,32 +12,26 @@
             secondaryIcon: secondaryIcon,
             disabled: disabled,
             customLabel: customLabel,
-            itemWidth: itemWidth,
-            childItems: childItems
+            itemWidth: itemWidth
         }
     "
     [ngClass]="{ 'fd-menu__list--separated': separated }"
 >
 </ng-container>
 <ng-template #itemTemplate let-label="label" let-i="i" let-active="active">
-    <li
-        data-tag="menu-item"
-        [attr.data-index]="i"
-        [attr.disabled]="disabled"
-        [ngClass]="['normal-item-state', disabled ? 'is-disabled' : '']"
-    >
+    <li data-tag="menu-item" tabindex="0" [attr.data-index]="i" [attr.disabled]="disabled">
         <!-- removed from li [ngClass]="{ 'fd-menu__list--separated': separated }" -->
         <div class="fd-menu__addon-before" *ngIf="icon; else noIcon">
-            <span [ngClass]="[icon]" data-tag="menu-item__icon-before" [attr.title]="icon">
+            <span [ngClass]="[icon]" data-tag="menu-item__icon" [attr.title]="icon">
                 <!-- removed from span , separated ? 'fd-menu__list--separated' : '' -->
             </span>
         </div>
         <div class="fd-menu__label-truncation">
             <span
                 class="fd-menu__item"
-                [ngClass]="[active ? 'is-selected' : '']"
+                [ngClass]="[active ? 'is-selected' : '', disabled ? 'is-disabled' : '']"
                 [ngStyle]="{ width: getItemWidth() }"
-                (click)="childItems.length > 0 ? doSomethingElse(childItems) : onItemClick()"
+                (click)="onItemClick()"
                 role="button"
                 data-tag="menu-item__button"
                 [attr.title]="customLabel ? customLabel : label"
@@ -48,7 +42,7 @@
         <div class="fd-menu__addon-after" *ngIf="secondaryIcon; else noIcon">
             <span
                 [ngClass]="[secondaryIcon ? secondaryIcon : '']"
-                data-tag="menu-item__icon-after"
+                data-tag="menu-item__icon"
                 [attr.title]="secondaryIcon"
             >
                 <!-- removed from span , separated ? 'fd-menu__list--separated' : '' -->
@@ -59,19 +53,13 @@
 <ng-template #noIcon><span [ngClass]="{ 'fd-menu__list--separated': separated }"></span></ng-template>
 <!-- works fine don't change -->
 <ng-template #selectableItemTemplate let-label="label" let-i="i" let-selected="selected" let-active="active">
-    <li
-        data-tag="menu-item"
-        [attr.data-index]="i"
-        [attr.disabled]="disabled"
-        [ngClass]="[selected ? 'selected-item-state' : 'normal-item-state', disabled ? 'is-disabled' : '']"
-        [attr.aria-selected]="selected ? 'true' : 'false'"
-    >
+    <li data-tag="menu-item" tabindex="0" [attr.data-index]="i" [attr.disabled]="disabled">
         <!-- removed from li         [ngClass]="{ 'fd-menu__list--separated': separated }" -->
         <div class="fd-menu__addon-before">
             <span
                 *ngIf="selected; else noIcon"
                 [ngClass]="[icon ? icon : 'sap-icon--accept']"
-                data-tag="menu-item__icon-before"
+                data-tag="menu-item__icon"
                 [attr.title]="icon"
             >
                 <!-- removed from span , separated ? 'fd-menu__list--separated' : '' -->
@@ -80,23 +68,15 @@
         <div class="fd-menu__label-truncation">
             <span
                 class="fd-menu__item"
-                [ngClass]="[active ? 'is-selected' : '']"
-                (click)="childItems.length > 0 ? doSomethingElse(childItems) : onItemClick()"
+                [ngClass]="[active ? 'is-selected' : '', disabled ? 'is-disabled' : '']"
+                (click)="onItemClick()"
                 [ngStyle]="{ width: getItemWidth() }"
+                [attr.aria-selected]="selected ? 'true' : 'false'"
                 role="button"
                 data-tag="menu-item__button"
                 [attr.title]="customLabel ? customLabel : label"
             >
                 {{ label }}
-            </span>
-        </div>
-        <div class="fd-menu__addon-after" *ngIf="secondaryIcon; else noIcon">
-            <span
-                [ngClass]="[secondaryIcon ? secondaryIcon : '']"
-                data-tag="menu-item__icon-after"
-                [attr.title]="secondaryIcon"
-            >
-                <!-- removed from span , separated ? 'fd-menu__list--separated' : '' -->
             </span>
         </div>
     </li>
