+++ libs/platform/src/lib/components/menu/menu-component/menu-item.component.ts
@@ -6,14 +6,10 @@ import {
     Output,
     EventEmitter,
     AfterViewInit,
-    ViewEncapsulation,
-    ComponentFactoryResolver,
-    ElementRef,
-    HostListener
+    ViewEncapsulation
 } from '@angular/core';
 import { MenuItem, MenuGroup } from './menu.component';
 import { Highlightable } from '@angular/cdk/a11y';
-import { DefaultMenuItem, MenuItemDirective, MenuKeyboardService } from '@fundamental-ngx/core';
 
 @Component({
     selector: 'fdp-menu-item',
@@ -21,55 +17,28 @@ import { DefaultMenuItem, MenuItemDirective, MenuKeyboardService } from '@fundam
     styleUrls: ['./menu-item.component.scss'],
     encapsulation: ViewEncapsulation.None
     // providers: [MenuKeyboardService]
-}) /*extends MenuItemDirective*/
-export class MenuItemComponent implements OnInit, OnDestroy, DefaultMenuItem, AfterViewInit, Highlightable {
-    @Input()
-    public label: string;
-    @Input()
-    public index: string;
-    @Input()
-    public icon: string;
-    @Input()
-    public selectable: boolean;
-    @Input()
-    public selected: boolean;
-    @Input()
-    public secondaryIcon: string;
+})
+export class MenuItemComponent implements OnInit, OnDestroy, /*DefaultMenuItem, */ AfterViewInit, Highlightable {
+    @Input() public label: string;
+    @Input() public index: string;
+    @Input() public icon: string;
+    @Input() public selectable: boolean;
+    @Input() public selected: boolean;
+    @Input() public secondaryIcon: string;
 
-    @Input()
-    public item: MenuItem;
-    @Input()
-    public group: MenuGroup;
+    @Input() public item: MenuItem;
+    @Input() public group: MenuGroup;
 
-    @Input()
-    public separated: boolean;
-    @Input()
-    public disabled: boolean;
-    @Input()
-    public customLabel: string;
-    @Input()
-    public itemWidth: string;
-
-    @Input()
-    public childItems: MenuItem[] = [];
+    @Input() public separated: boolean;
+    @Input() public disabled: boolean;
+    @Input() public customLabel: string;
+    @Input() public itemWidth: string;
 
     public _isActive = false;
 
-    currentAdIndex = -1;
-
     @Output() itemClick: EventEmitter<void> = new EventEmitter();
 
-    /**  Event thrown, when there is some keyboard event detected on mega menu item */
-    @Output()
-    readonly keyDown: EventEmitter<KeyboardEvent> = new EventEmitter<KeyboardEvent>();
-
-    constructor(
-        private componentFactoryResolver: ComponentFactoryResolver,
-        private itemEl: ElementRef,
-        private menuKeyboardService: MenuKeyboardService
-    ) {
-        // super(itemEl);
-    }
+    constructor() {}
 
     ngOnInit() {
         // this._isActive = this.selected ? true : false;
@@ -90,61 +59,17 @@ export class MenuItemComponent implements OnInit, OnDestroy, DefaultMenuItem, Af
     }
 
     onItemClick() {
-        console.log('child items ' + this.childItems.length);
         this.itemClick.emit();
     }
 
     getItemWidth(): string {
-        // todo: handle em, rem etc.
-        let finalItemWidth: string = '';
-        if (this.itemWidth !== undefined) {
-            const itemWidthNumber: number = Number(this.itemWidth.split('px')[0]);
-            finalItemWidth = this.itemWidth;
-            if (this.item.secondaryIcon !== undefined || this.item.secondaryIcon !== '') {
-                // secondary icon exists
-                finalItemWidth = itemWidthNumber - 85 + 'px';
-            }
+        const itemWidthNumber: number = Number(this.itemWidth.split('px')[0]);
+        let finalItemWidth: string = this.itemWidth;
+        if (this.item.secondaryIcon !== undefined || this.item.secondaryIcon !== '') {
+            // secondary icon exists
+            finalItemWidth = itemWidthNumber - 85 + 'px';
         }
-        // console.log(finalItemWidth + 'is finalItemWidth and itemWidth ' + itemWidthNumber);
+        console.log(finalItemWidth + 'is finalItemWidth and itemWidth ' + itemWidthNumber);
         return finalItemWidth;
-    }
-
-    doSomethingElse(data: MenuItem[]) {
-        // alert('open new menu ');
-        // this.currentAdIndex = (this.currentAdIndex + 1) % this.ads.length;
-        // const adItem = this.childItems[this.currentAdIndex];
-        // const componentFactory = this.componentFactoryResolver.resolveComponentFactory(adItem.component);
-
-        // const viewContainerRef = this.data.viewContainerRef;
-        // viewContainerRef.clear();
-
-        // for now do item click only
-        this.itemClick.emit();
-    }
-
-    public focus(): void {
-        console.log('calling focussss');
-        // if (this.group.label) {
-        //     this.itemEl.nativeElement.group.groupItems[0].focus();
-        // }
-        this.itemEl.nativeElement.focus();
-    }
-
-    public click(): void {
-        this.itemClick.emit();
-    }
-
-    /** @hidden */
-    @HostListener('keydown', ['$event'])
-    handleKeyboardEvent(event: KeyboardEvent) {
-        console.log('@@@@@@@ event code in item ' + event.code);
-        switch (event.code) {
-            case 'Space':
-            case 'Enter':
-                this.itemClick.emit();
-                break;
-            default:
-                this.keyDown.emit(event);
-        }
     }
 }
