+++ libs/platform/src/lib/components/menu/menu-component/menu.component.scss
@@ -36,22 +36,7 @@
     border-bottom-style: solid;
     border-bottom-width: 1px;
     color: var(--fd-menu-item-separator-color);
-    &:focus {
-        //selected state
-        outline: var(--fd-color-action-focus) dotted 1px;
-        background-color: rgba(10, 110, 209, 0.07);
-        // background-color: var(--fd-color-action-hover);
-    }
     // &:first-child {
     //     border-top-width: 0;
     // }
-}
-
-// custom scrollbar: WILL NOT WORK ON FIREFOX, IE and EDGE
-::-webkit-scrollbar {
-    width: 4px;
-    height: 80px;
-}
-::-webkit-scrollbar-thumb {
-    background: #89919a;
 }
