+++ libs/platform/src/lib/components/menu/menu-component/menu.component.ts
@@ -1,24 +1,22 @@
 import {
-    AfterViewInit,
-    ChangeDetectionStrategy,
-    ChangeDetectorRef,
     Component,
-    ElementRef,
-    Input,
-    OnChanges,
+    ChangeDetectorRef,
     OnInit,
+    OnDestroy,
+    OnChanges,
     SimpleChanges,
+    Input,
     ViewEncapsulation,
-    ViewChildren,
+    ChangeDetectionStrategy,
+    ElementRef,
     QueryList,
-    AfterContentInit,
-    ContentChildren,
-    OnDestroy
+    ViewChildren,
+    AfterViewInit
 } from '@angular/core';
+import { ActiveDescendantKeyManager } from '@angular/cdk/a11y';
 import { MenuItemComponent } from './menu-item.component';
-import { MenuKeyboardService } from '@fundamental-ngx/core';
-import { Subject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
+import { UP_ARROW, DOWN_ARROW, ENTER } from '@angular/cdk/keycodes';
+import { group } from '@angular/animations';
 
 /**
  * Interface that represents menu item.
@@ -64,14 +62,10 @@ export interface MenuItem {
      */
     customLabel?: string;
 
-    childItems?: MenuItem[]; // move this to MegaMenu
-
-    id?: number;
-
     /**
      * Callback function which will execute when menu item is clicked.
      */
-    command(): void;
+    callback(): void;
 }
 
 /**
@@ -102,11 +96,11 @@ export interface MenuGroup {
     /**
      * List of menu items of the group.
      */
-    groupItems: MenuItem[];
+    children: MenuItem[];
 }
 
 /**
- * `<fdp-menu>` is a menu component which provides navigation, action and selection
+ * `<af-menu>` is a menu component which provides navigation, action and selection
  * options.
  *
  * ```html
@@ -119,12 +113,12 @@ export interface MenuGroup {
  * var menu = document.getElementByTagName('fdp-menu')[0];
  * var data = [{
  *   label: 'Item One',
- *   command: () => {
+ *   callback: () => {
  *     alert('The first item.')
  *   }
  * }, {
  *   label: 'Item Two',
- *   command: () => {
+ *   callback: () => {
  *     alert('The second item.')
  *   }
  * }];
@@ -143,14 +137,11 @@ export interface MenuGroup {
     changeDetection: ChangeDetectionStrategy.OnPush,
     encapsulation: ViewEncapsulation.None
 })
-// AfterContentInit, OnDestroy
-// AfterViewInit
-export class MenuComponent implements OnInit, OnChanges, AfterViewInit, AfterContentInit, OnDestroy {
+export class MenuComponent implements OnInit, OnChanges, AfterViewInit, OnDestroy {
     /**
      * Add separating line between menu items. [Default: false]
      */
-    @Input()
-    public showSeparator = false;
+    @Input() public separator = false;
 
     /**
      * Display menu groups as columns. [Default: false]
@@ -161,88 +152,96 @@ export class MenuComponent implements OnInit, OnChanges, AfterViewInit, AfterCon
      * Alignment of menu items; either "left", "right" or "inherit".
      * [Default "inherit"]
      */
-    @Input()
-    public textAlign: string;
+    @Input() public textAlign: string;
 
     /**
      * Specify if items are to scroll
      */
-    @Input()
-    public isScrolling: boolean = false;
+    @Input() public scrolling: boolean = false;
 
     /**
      * Specify the number of items after which scroll must begin
      */
-    @Input()
-    public scrollLimit: number;
+    @Input() public scrollLimit: number;
 
     /**
      * the max width of the menu component
      */
-    @Input()
-    public width: string;
-
-    /**
-     * The menu items that are passed in by the user.
-     */
-    @Input()
-    public menuItems: (MenuItem | MenuGroup)[] = [];
+    @Input() public width: string;
 
     public groups: MenuGroup[];
     private target: ElementRef;
     private isIcon: boolean = false;
     private numberOfItems: number = 0;
-    private data = [];
 
-    @ViewChildren(MenuItemComponent)
-    menuQueryList: QueryList<MenuItemComponent>;
+    @ViewChildren(MenuItemComponent) public menuItems: QueryList<MenuItemComponent>;
+    private keyManager: ActiveDescendantKeyManager<MenuItemComponent>;
 
     /** An RxJS Subject that will kill the data stream upon component’s destruction (for unsubscribing)  */
-    private readonly onDestroy$: Subject<void> = new Subject<void>();
+    // private readonly onDestroy$: Subject<void> = new Subject<void>();
 
-    constructor(private cd: ChangeDetectorRef, private keyboardService: MenuKeyboardService) {}
+    /**
+     * Load menu data.
+     * @param data Data can be either array of MenuItems or array of MenuGroups.
+     */
+    @Input() load = (data: (MenuItem | MenuGroup)[]) => {
+        this.groups = this.processData(data);
+        this.cd.markForCheck();
+    };
+
+    /**
+     * Registers 'keyup' event listener on provided element.
+     */
+    @Input() registerKeyupListener = (elem: ElementRef) => {
+        if (this.target) {
+            this.target.nativeElement.removeEventListener('keyup', this.onTargetKeyup);
+        }
+        this.target = elem;
+        this.target.nativeElement.addEventListener('keyup', this.onTargetKeyup.bind(this));
+    };
+
+    initializeKeyManager(): void {
+        this.menuItems.forEach(item => {
+            // remove all focused items
+            console.log(' some item ');
+            console.log(item);
+            item.setInactiveStyles();
+        });
+        //   this.menuItems.forEach(itemObj => {
+        //     // remove all focused item
+        //     itemObj.group.children.forEach(item =>{
+        //       console.log(' some item ' + item.label);
+        //       item.setInactiveStyles();
+        //     })
+
+        // });
+        this.keyManager = new ActiveDescendantKeyManager(this.menuItems).withWrap(false);
+    }
+
+    constructor(private cd: ChangeDetectorRef) {}
 
     ngOnInit() {
-        if ((this.isScrolling && this.scrollLimit === undefined) || (!this.isScrolling && this.scrollLimit > 0)) {
-            // if scroll limit was not specified but isScrolling flag was used, use default value
-            // or if scroll limit was specified but isScrolling flag was not marked true, even then use default value
+        if ((this.scrolling && this.scrollLimit === undefined) || (!this.scrolling && this.scrollLimit > 0)) {
+            // if scroll limit was not specified but scrolling flag was used, use default value
+            // or if scroll limit was specified but scrolling flag was not marked true, even then use default value
             this.scrollLimit = this.numberOfItems; // default
-            // can also disable scrolling explicitly
-            this.isScrolling = false;
         }
-        this.groups = this.processData(this.menuItems);
-        console.log('group is ' + this.groups.length);
-        this.cd.markForCheck();
     }
 
     ngOnChanges(changes: SimpleChanges) {
-        // this.cd.detectChanges();
-        this.cd.markForCheck();
+        this.cd.detectChanges();
     }
 
     ngAfterViewInit() {
-        this.data = this.menuQueryList.toArray();
-        this.data.sort((a, b) => {
-            console.log('a lbel: ' + a.group.groupItems + ' at index' + a.item.id);
-            // console.log(a);
-            // console.log(b);
-            console.log('b lbel: ' + b.item.label + ' at index' + b.item.id);
-            // return a.item.id > b.item.id ? a.item.id : b.item.id;
-            return a.index - b.index;
-        });
-        console.log('after sort');
-        console.log(this.data);
-        // this.cd.markForCheck();
-        // console.log(this.menuQueryList.toArray());
+        console.log('in menu ' + this.menuItems.length);
+        this.initializeKeyManager();
+        this.cd.markForCheck();
+    }
 
-        this.data.forEach((item: MenuItemComponent, index: number) =>
-            item.keyDown.pipe(takeUntil(this.onDestroy$)).subscribe((keyboardEvent: KeyboardEvent) => {
-                console.log('item index is ' + item.index + ' and index ' + index);
-
-                this.handleKeyPress(keyboardEvent, parseInt(item.index, 10));
-            })
-        );
-        // this.cd.markForCheck();
+    ngOnDestroy(): void {
+        if (this.target) {
+            this.target.nativeElement.removeEventListener('keyup', this.onTargetKeyup);
+        }
     }
 
     onItemClick(item: MenuItem, _group: MenuGroup): void {
@@ -255,140 +254,90 @@ export class MenuComponent implements OnInit, OnChanges, AfterViewInit, AfterCon
             item.selected = !item.selected;
             this.cd.markForCheck();
         }
-        item.command();
+        item.callback();
     }
 
     isMenuGroup(item: MenuItem | MenuGroup): item is MenuGroup {
-        return (item as MenuGroup).groupItems !== undefined;
+        return (item as MenuGroup).children !== undefined;
     }
 
     processData(data: (MenuItem | MenuGroup)[]): MenuGroup[] {
         const groups: MenuGroup[] = [];
         let newGroup: MenuItem[] = [];
 
-        let index = 0;
-        if (data !== undefined) {
-            data.forEach(record => {
-                console.log(record.label + 'is the label before pushing');
-                if (this.isMenuGroup(record)) {
-                    if (newGroup.length > 0) {
-                        console.log('adding previously unpushed individual items into group @@@@@@@@@');
-
-                        this.numberOfItems++;
-                        // newGroup.forEach(groupItem => {
-                        //     groupItem.id = index;
-                        //     index++;
-                        // });
-                        groups.push({
-                            groupItems: newGroup
-                        });
-                        // console.log('group 1 ');
-                        // console.log(groups);
-                        newGroup = [];
-                    }
-                    console.log('adding a new group with header @@@@@@@@@');
-
+        data.forEach(record => {
+            console.log(record.label + 'is the label before pushing');
+            if (this.isMenuGroup(record)) {
+                if (newGroup.length > 0) {
                     this.numberOfItems++;
-                    record.groupItems.forEach(groupItem => {
-                        groupItem.id = index;
-                        index++;
+                    groups.push({
+                        children: newGroup
                     });
-                    groups.push(record);
-                    // console.log('group 2 ');
-                    // console.log(groups);
-                } else {
-                    console.log('no header, adding individual items @@@@@@@@@');
-
-                    this.numberOfItems++;
-                    record.id = index;
-                    index++;
-                    newGroup.push(record);
-                    // console.log('group 3 ' + record.id);
-                    // console.log(groups);
+                    console.log('group 1 ');
+                    console.log(groups);
+                    newGroup = [];
                 }
-            });
-        }
-        // if no group headers present or if last set of items have not yet been pushed to a group
+                this.numberOfItems++;
+                groups.push(record);
+                console.log('group 2 ');
+                console.log(groups);
+            } else {
+                this.numberOfItems++;
+                newGroup.push(record);
+                console.log('group 3 ');
+                console.log(groups);
+            }
+        });
+        // if no group headers present
         if (newGroup.length > 0) {
-            console.log('what does this else do? @@@@@@@@@');
-
             // this.numberOfItems++;  //todo : check if this is needed. maybe needed.
-            // newGroup.forEach(groupItem => {
-            //     groupItem.id = index;
-            //     index++;
-            // });
             groups.push({
-                groupItems: newGroup
+                children: newGroup
             });
-            // console.log('group 4 ');
+            console.log('group 4 ');
+            console.log(groups);
         }
-        console.log(groups);
         return groups;
     }
 
-    // isIconPresent() {
-    //     for (const _group of this.groups) {
-    //         if (_group.icon) {
-    //             // console.log('group icon: ' + group.icon);
-    //             this.isIcon = true;
-    //             break;
-    //         } else if (_group.groupItems) {
-    //             for (const item of _group.groupItems) {
-    //                 if (item.icon) {
-    //                     // console.log('item icon: ' + item.icon);
-    //                     this.isIcon = true;
-    //                     break;
-    //                 }
-    //             }
-    //         } else {
-    //             this.isIcon = false;
-    //             break;
-    //         }
-    //     }
-    //     return this.isIcon;
-    // }
-
-    /** @hidden */
-    ngAfterContentInit(): void {
-        // this.menuQueryList.forEach((item: MenuItemComponent, index: number) =>
-        //     item.keyDown
-        //         .pipe(takeUntil(this.onDestroy$))
-        //         .subscribe((keyboardEvent: KeyboardEvent) => this.handleKeyPress(keyboardEvent, index))
-        // );
+    isIconPresent() {
+        for (const _group of this.groups) {
+            if (_group.icon) {
+                // console.log('group icon: ' + group.icon);
+                this.isIcon = true;
+                break;
+            } else if (_group.children) {
+                for (const item of _group.children) {
+                    if (item.icon) {
+                        // console.log('item icon: ' + item.icon);
+                        this.isIcon = true;
+                        break;
+                    }
+                }
+            } else {
+                this.isIcon = false;
+                break;
+            }
+        }
+        return this.isIcon;
+        // return true;
     }
 
-    handleKeyPress(event: KeyboardEvent, index: number) {
-        console.log('yo yo handling at ' + index + 'for event ' + event.key);
-        // console.log(this.menuQueryList);
-        // const data = this.menuQueryList.toArray();
-        // data.sort((a, b) => {
-        //     console.log('a lbel: ' + a.item.label + ' at index' + a.index);
-        //     console.log('b lbel: ' + b.item.label + ' at index' + b.index);
-        //     return a.item.label.localeCompare(b.item.label);
-        // });
-        // console.log('after sort');
-        // console.log(data);
-        // if(event.key === 'ArrowDown'){
-
-        // }else{
-        // this.keyboardService.keyDownHandler(event, index, this.menuQueryList.toArray());
-        this.keyboardService.keyDownHandler(event, index, this.data);
-        // }
-    }
-
-    focusFirst() {
-        console.log('menu query list:--------');
-
-        console.log(this.menuQueryList);
-
-        console.log('end--------');
-
-        setTimeout(() => this.menuQueryList.first.focus(), 0);
-    }
-
-    /** @hidden */
-    ngOnDestroy(): void {
-        this.onDestroy$.next();
-        this.onDestroy$.complete();
+    onTargetKeyup($event: any) {
+        console.log('$event keycode ' + $event.keyCode);
+        if (this.keyManager.activeItem) {
+            console.log('item is ');
+            console.log(this.keyManager.activeItem);
+        }
+        if ($event.keyCode === UP_ARROW || $event.keyCode === DOWN_ARROW) {
+            this.keyManager.onKeydown($event);
+        } else if ($event.keyCode === ENTER) {
+            const activeItem = this.keyManager.activeItem;
+            if (activeItem) {
+                this.onItemClick(activeItem.item, activeItem.group);
+                this.initializeKeyManager();
+            }
+        }
+        this.cd.markForCheck();
     }
 }
